[migrate-mongo] Fix FalseCJS: split index.d.ts into ESM/CJS twins with matching exports
migrate-mongo ships dual ESM/CJS builds via package.json exports conditions,
but the types package had a single ambient index.d.ts with no exports map,
so it resolved as CJS-only under node16-from-ESM.
