+++ types/migrate-mongo/index.d.ts
@@ -1,117 +1,1 @@
-/// <reference types="node" />
-
-import * as mongo from "mongodb";
-
-export function init(): Promise<void>;
-export function create(description: string): Promise<string>;
-export namespace database {
-    function connect(): Promise<{
-        client: mongo.MongoClient;
-        db: mongo.Db & { close: mongo.MongoClient["close"] };
-    }>;
-}
-export namespace config {
-    /**
-     * @internal
-     */
-    const DEFAULT_CONFIG_FILE_NAME: string;
-    /**
-     * @internal
-     */
-    function shouldExist(): Promise<void>;
-    /**
-     * @internal
-     */
-    function shouldNotExist(): Promise<void>;
-    /**
-     * @internal
-     */
-    function getConfigFilename(): string;
-    /**
-     * Read the `migrate-mongo-config.js` file.
-     */
-    function read(): Promise<Config>;
-    /**
-     * Set the passed config object.
-     * @param config The config object.
-     */
-    function set(config: Partial<Config>): void;
-
-    interface Config {
-        mongodb: {
-            url: Parameters<typeof mongo.MongoClient["connect"]>[0];
-            databaseName?: mongo.Db["databaseName"];
-            options?: mongo.MongoClientOptions;
-        };
-        /**
-         * The migrations dir, can be an relative or absolute path.
-         */
-        migrationsDir?: string | undefined;
-        /**
-         * The MongoDB collection where the applied changes are stored.
-         */
-        changelogCollectionName: string;
-        /**
-         * The file extension to create migrations and search for in migration dir
-         */
-        migrationFileExtension?: string | undefined;
-        /**
-         * Enable the algorithm to create a checksum of the file contents and use that in the comparison to determine
-         * if the file should be run.  Requires that scripts are coded to be run multiple times.
-         */
-        useFileHash?: boolean | undefined;
-        /**
-         * The MongoDB collection where the lock will be created.
-         */
-        lockCollectionName?: string | undefined;
-        /**
-         * The value in seconds for the TTL index that will be used for the lock. Value of 0 will disable the feature.
-         */
-        lockTtl?: number | undefined;
-        /**
-         * The module system used by the config file and migration scripts.
-         */
-        moduleSystem?: "commonjs" | "esm" | undefined;
-    }
-}
-
-/**
- * Apply all pending migrations.
- * @example
- * ```js
- * const db = await database.connect();
- * const migrated = await up(db);
- * migrated.forEach(fileName => console.log('Migrated:', fileName));
- * ```
- * If an an error occurred, the promise will reject and won't continue with the rest of the pending migrations.
- */
-export function up(db: mongo.Db, client: mongo.MongoClient): Promise<string[]>;
-/**
- * Revert (only) the last applied migration.
- * @example
- * ```js
- * const db = await database.connect();
- * const migratedDown = await down(db);
- * migratedDown.forEach(fileName => console.log('Migrated Down:', fileName));
- * ```
- */
-export function down(db: mongo.Db, client: mongo.MongoClient): Promise<string[]>;
-export function status(db: mongo.Db): Promise<MigrationStatus[]>;
-
-export interface MigrationStatus {
-    fileName: string;
-    fileHash?: string | undefined;
-    /**
-     * Either "PENDING" or a JSON date.
-     */
-    appliedAt: string;
-    /**
-     * Timestamp shared by migrations that were applied together in the same `up`/`down` run.
-     */
-    migrationBlock?: number | undefined;
-}
-
-/**
- * Type of `up()` and `down()` functions for migration scripts.
- */
-export type MigrationFunction = (db: mongo.Db, client: mongo.MongoClient) => Promise<void>;
+export * from "./index.cjs";
